backend/requirements.txt
# LLM runtime
llama-cpp-python==0.2.53

# Cassandra
cassandra-driver==3.29.0

# API
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3

# Утилиты
orjson==3.10.1
loguru==0.7.2
tqdm==4.66.2
python-dotenv==1.0.1